+++ src/Vertex.java
@@ -2,8 +2,8 @@ import processing.core.PVector;
 
 
 public class Vertex {
-	public int x, y, z, dx, dy, dz;
-	public Vertex( int x_, int y_, int z_, int dx_, int dy_, int dz_ ) {
+	public float x, y, z, dx, dy, dz;
+	public Vertex( float x_, float y_, float z_, float dx_, float dy_, float dz_ ) {
 		this.x = x_;
 		this.y = y_;
 		this.z = z_;
